+++ src/index.ts
@@ -2,4 +2,4 @@ export * from './types';
 export * from './decorators';
 export * from './restBase';
 export * from './restRoute';
-export { RestModel } from './restModel';
+export * from './restModel';