+++ src/restBase.ts
@@ -1,6 +1,7 @@
 import { HttpClient, HttpRequest, HttpEvent } from '@angular/common/http';
 import { Observable } from 'rxjs';
 import { RestRoute } from './restRoute';
+import { HttpMethod } from './types';
 
 export class RestBase {
   constructor (protected http: HttpClient) {
@@ -10,15 +11,11 @@ export class RestBase {
     return req;
   }
 
-  protected responseInterceptor(res: Observable<any>): Observable<any> {
+  protected responseInterceptor(res: Observable<HttpEvent<any>>): Observable<HttpEvent<any>> {
     return res;
   }
 
-  protected FullResponseInterceptor(res: Observable<HttpEvent<any>>): Observable<HttpEvent<any>> {
-    return res;
-  }
-
-  protected mapModel(path: string, data: any) {
+  protected mapModel(method: HttpMethod, path: string, data: any) {
     return data;
   }