+++ src/restRoute.ts
@@ -1,77 +1,84 @@
-import { HttpRequest, HttpParams, HttpHeaders, HttpEventType } from '@angular/common/http';
-import { Resource, IAbstractBase } from './types';
-import { RestModel, RestModelBase } from './restModel';
+import { HttpRequest, HttpParams, HttpHeaders, HttpEventType, HttpResponse } from '@angular/common/http';
+import { Resource, IAbstractBase, HttpMethod } from './types';
+import { RestModel, RestModelBase } from './index';
 import { Observable } from 'rxjs';
-import { map } from 'rxjs/operators';
+import { map, filter } from 'rxjs/operators';
 
 export class RestRoute {
-  constructor (private base: IAbstractBase, private path: string) {
+  private _path?: string;
+
+  constructor (protected _base: IAbstractBase, _path?: string) {
+    if (_path) {
+      this._path = _path;
+    }
   }
 
-  private createHttpRequest(method: 'GET'|'POST', params?: HttpParams, id_data?: any) {
+  protected createHttpRequest(req: HttpRequest<any>) {
+    // pass through request interceptor
+    req = this._base.requestInterceptor(req);
+
+    // the observable to return
+    let observable = this._base.http.request<any>(req);
+
+    // pass through response interceptor
+    observable = this._base.responseInterceptor(observable);
+
+    return observable.pipe(
+      filter(response => response.type === HttpEventType.Response),
+      map((response: HttpResponse<any>) => response.body)
+    );
+  }
+
+  private createRouteHttpRequest(method: 'GET'|'POST', params?: HttpParams, id_data?: any) {
     const url = this.getFullPath(method === 'GET' ? id_data : null);
     const headers = new HttpHeaders(this.getDefaultHeaders());
 
-    let req = new HttpRequest(method, url,
+    const req = new HttpRequest(method, url,
       method === 'POST' ? id_data : null,
       { headers, params }
     );
 
-    // pass through request interceptor
-    req = this.base.requestInterceptor(req);
-
-    // the observable to return
-    const requestObservable = this.base.http.request<any>(req);
-
-    const observable = new Observable<any>(observer => {
-      // pass through response interceptor
-      this.base
-        .FullResponseInterceptor(requestObservable)
-        .subscribe(response => {
-          if (response.type === HttpEventType.Response) {
-            observer.next(response.body);
-          }
-        });
-    });
-
-    return observable;
+    return this.createHttpRequest(req);
   }
 
-  private makeRest<T>(data: any): RestModel<T> {
-    const model = this.mapModel(this.path, data);
+  protected makeRest<T>(method: HttpMethod, data: any): RestModel<T> {
+    const model = this.mapModel(method, this._path || this._base.resource.path, data);
 
-    const base = { ...this.base };
-    const proto = Object.getPrototypeOf(this.base);
+    const base = { ...this._base };
+    const proto = Object.getPrototypeOf(this._base);
     Object.setPrototypeOf(base, proto);
 
-    const resource: Resource = {
-      id: data.id,
-      path: this.path,
-      parent: base.resource
-    };
+    if (this._path) {
+      const resource: Resource = {
+        id: data.id,
+        path: this._path,
+        parent: base.resource
+      };
 
-    base.resource = resource;
+      base.resource = resource;
+    }
+
     return new RestModelBase<T>(base as any, model) as any;
   }
 
-  getList<T>(params?: HttpParams | undefined): Observable<Array<RestModel<T>>> {
-    return this.createHttpRequest('GET', params)
-      .pipe(map((response: any[]) => response.map(r => this.makeRest<T>(r))));
+  getList<T>(params?: HttpParams): Observable<Array<RestModel<T>>> {
+    return this.createRouteHttpRequest('GET', params)
+      .pipe(map((response: any[]) => response.map(r => this.makeRest<T>(HttpMethod.GET, r))));
   }
 
-  getOne<T>(id: number, params?: HttpParams | undefined): Observable<RestModel<T>> {
-    return this.createHttpRequest('GET', params, id)
-      .pipe(map(response => this.makeRest<T>(response)));
+  getOne<T>(id: number, params?: HttpParams): Observable<RestModel<T>> {
+    return this.createRouteHttpRequest('GET', params, id)
+      .pipe(map(response => this.makeRest<T>(HttpMethod.GET, response)));
   }
 
-  post<T>(data: any, params?: HttpParams | undefined): Observable<RestModel<T>> {
-    return this.createHttpRequest('POST', params, data)
-      .pipe(map(response => this.makeRest<T>(response)));
+  post<T>(data: any, params?: HttpParams): Observable<RestModel<T>> {
+    return this.createRouteHttpRequest('POST', params, data)
+      .pipe(map(response => this.makeRest<T>(HttpMethod.POST, response)));
   }
 
-  private getFullPath(id?: number) {
+  protected getFullPath(id?: number) {
     let parentUrl = '/';
-    addRoute(this.base.resource);
+    addRoute(this._base.resource);
 
     function addRoute(route?: Resource) {
       if (route) {
@@ -91,20 +98,25 @@ export class RestRoute {
       baseurl = baseurl.slice(0, -1);
     }
 
-    return baseurl + parentUrl + this.path + (id ?  '/' + id : '');
+    if (!this._path) {
+      parentUrl = parentUrl.slice(0, parentUrl.length-1);
+    }
+
+    const url = baseurl + parentUrl + (this._path || '') + (id ?  '/' + id : '');
+    return url;
   }
 
   // Base
 
-  private getBaseUrl(): string {
-    return this.base.getBaseUrl();
+  protected getBaseUrl(): string {
+    return this._base.getBaseUrl();
   }
 
-  private getDefaultHeaders(): { [header: string]: string | string[]; } {
-    return this.base.getDefaultHeaders();
+  protected getDefaultHeaders(): { [header: string]: string | string[]; } {
+    return this._base.getDefaultHeaders();
   }
 
-  private mapModel(path: string, data: any): any {
-    return this.base.mapModel(path, data);
+  protected mapModel(method: HttpMethod, path: string, data: any): any {
+    return this._base.mapModel(method, path, data);
   }
 }